+++ src/components/Card/CardHeader.tsx
@@ -6,7 +6,7 @@ import CardTitle from "./CardTitle";
 import CardOptions from "./CardOptions";
 
 interface CardHeaderProps extends OverviewItemProps {
-    children?: any; // Fixme: workaround to prevent typescript problem with child.type
+    children: JSX.Element | (JSX.Element | undefined | null)[] | null | undefined;
 }
 
 function CardHeader({
@@ -18,8 +18,8 @@ function CardHeader({
     let actions: any[] = [];
     let description: any[] = [];
 
-    React.Children.map(children, (child, i) => {
-        if (typeof child === "object" && !!child && !!child.type) {
+    children && (Array.isArray(children) ? children : [children]).forEach((child) => {
+        if (typeof child === "object" && child && !!child.type) {
             switch (child.type) {
                 case CardTitle:
                     description.push(child);